[android] Search by categories UI.

+++ android/jni/com/mapswithme/core/jni_helper.hpp
@@ -53,8 +53,8 @@ jobject GetNewPoint(JNIEnv * env, m2::PointI const & point);
 template<typename TValue, typename TToJavaFn>
 jobjectArray ToJavaArray(JNIEnv * env, jclass clazz, vector<TValue> const & src, TToJavaFn && toJavaFn)
 {
-  int const size = src.size();
-  jobjectArray jArray = env->NewObjectArray(size, clazz, 0);
+  size_t const size = src.size();
+  jobjectArray jArray = env->NewObjectArray((jint) size, clazz, 0);
   for (size_t i = 0; i < size; i++)
   {
     TScopedLocalRef jItem(env, toJavaFn(env, src[i]));
